added user validation schema
