+++ db.json
@@ -1048,7 +1048,7 @@
   ],
   "users": [
     {
-      "first_name": "Andreea",
+      "first_name": "Andrei",
       "last_name": "Mihu",
       "id": 1005,
       "email": "Mihai_Mihu@gmail.com",
@@ -1059,7 +1059,7 @@
     },
     {
       "first_name": "Anastasia",
-      "last_name": "Mirela",
+      "last_name": "Mirel",
       "id": 1006,
       "email": "Anastasia_Mirela@gmail.com",
       "password": "yuihjkbnm",
